Combined OpenGL, GLES2, and Metal abstraction layer
With the deprecation of OpenGL in MacOS, NanoGUI requires a new
mechanism to render both GUI and 3D content in a cross-platform manner.
This commit introduces a set of cross-API abstractions for textures,
render targets, shaders, uniform and varying shader parameters, and
render passes that works on OpenGL, GLES 2, and Metal.