update pybind11, switch to regular pybind11_mkdoc package
pybind11 no longer includes the 'mkdoc' script, which was moved to a
separate repository. This now depends on the 'clang' PyPI package and
includes a few other improvements that make it easier to use across
platforms.

By switching to this variant, we unfortunately lose some customization
made by Stephen. Will need to revisit this at some point to see if
some of this should go into the 'pybind11_mkdoc' repository, but this
at least allows me to continue generating docstrings on my machine for
the meantime.