Enable HDR on Mac
On Mac (with Metal), try to enable HDR when the application requests a
float buffer.

If the current display mode does not support HDR, the system falls back
to SDR. This can happen in the following circumstances:

- The hardware does not support HDR.
- The application is in windowed mode and HDR is disabled in the user's
System Preferences.
- The application is in fullscreen mode. For fullscreen, the
application seems to invariably get an SDR display mode, probably due
to limitations of the GLFW back end (an area for future work).