Avoid async deadlock
nanogui::async schedules a function to be executed in the next main loop
iteration. It’s doing it by appending the function to a list of
functions, while holding a lock (for thread safety).

In the UI thread nanogui will process each of these functions one by
one, while holding the same lock. This means that while it’s executing
an async function, any thread calling `async` will be blocked, which can
easily lead to deadlocks.

To fix this, in the UI thread the lock is held only until the functions
are moved to a temporary storage. This way when the functions are
executed, async can still be called (and the function will be executed
in the *next* main loop iteration)