Added “inside” popups
Before popup windows always popped up at the side of the parent window.
This worked for smaller windows, but in case the window was too big (a
common example is if the window covers the whole screen), the popup was
shown outside of the screen.

This patch adds two extra constants: LeftInside and RightInside.
If these constants are used (instead of Left and Right), the popup will
be shown next to the widget, and not next to the parent window.