Rewrite nanogui run loop logic
NanoGUI previously used an ad-hoc scheme for refreshing the window
contents using a background thread that sent periodic redraw events. This
unfortunately meant that redraw events were not guaranteed to be aligned
with the screen's refresh rate.

This commits replaces the ``nanogui::runloop()`` function with the
function ``nanogui::run(RunMode run_mode)``, where ``run_mode`` is one
of the following options:

- ``RunMode::VSync``: refresh at VSync frequency, and avoid screen
  tearing.

- ``RunMode::Lazy``: only refresh when absolutely needed (e.g., due
  to mouse/keyboard events)

- ``RunMode::Eager``: run redraw loop in brute force mode without
  ever waiting (will use 100% CPU).

This setting can be adjusted later on via ``set_run_mode()``, e.g.,
to conserve energy when the application loses focus.