``nanogui::Screen``: Fractional mouse motion events
NanoGUI uses integer-based coordinates throughout the codebase. However,
on high-DPI screens, mouse motion events can be _fractional_ due to the
presence of scaling.

This commit adds a fractional mouse motion event to the ``Screen``
class, which can be useful to preserve smoother motion, e.g., in the
context of 3D trackball-style navigation.